add missing types
